backend/pyproject.toml
[project]
name = "book-data-integration-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "ruff>=0.12.5",
    "sqlacodegen>=3.0.0",
]